requirements.txt
seaborn
scipy
numpy
pandas
matplotlib
pyserial
